pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
description = ""
authors = ["Justin Wiley <justin.wiley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-community = "^0.0.34"
openai = "^1.23.6"
python-dotenv = "^1.0.1"
pre-commit = "^3.7.0"
dspy-ai = "^2.4.9"
instructor = "^1.3.4"
pydantic = "^2.8.2"
mistralai = "^0.4.2"
darkslide = "^6.0.0"
watchdog = "^4.0.2"
transformers = "^4.44.2"
huggingface-hub = "^0.24.6"
llama-index-llms-openai = "^0.2.7"
streamlit = "^1.38.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
detect-secrets = "^1.5.0"
black = {extras = ["jupyter"], version = "^24.4.2"}
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
